Convert case ref from int to long

+++ src/main/java/uk/gov/ons/census/casesvc/model/dto/PrintCaseSelected.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class PrintCaseSelected {
-  private int caseRef;
+  private long caseRef;
   private String packCode;
   private String actionRuleId;
   private String batchId;
